fix: enable creation of systemd-enabled container
Added centos.systemd.test.Dockerfile to create a SystemD-enabled
container, similar to [1].
This container seems to have access to systemD bus if run with the
command supplied in run_test.sh.
The official image centos/systemd:latest appears to function similarly
so this recreation is considered successful.
Variable file centos.systemd.test.vars was created to hold variables
specific to this container.
This file can be sourced prior to building or running a container, but
the Dockerfile callout is the only important variable defined therein.

[1] https://hub.docker.com/r/centos/systemd/