enable creation of servers with wait between
In any node connected to the openstack cluster managment network, the
create servers script can be run and it will use the openstack CLI to
create servers with a single port on the private network then wait for
them to respond to pings before continuing.
Service configuration was modified to reflect a pre-stacked image.