feat: enable closed-loop creation of tenants & net
"create and wait" can now create networks, subnets, and then instances
on them based on the identifiers of the tenants, compute hosts, etc.
Current implementation uses COMP_ID % NUM_NETWORKS to assign an instance
to a particular network.
Networks (and subnets) and instances that already exist will not be
re-created or created in duplicate.
Fixed openstackrc for host 20.
Existing instances may be pinged or not (VALIDATE_EXISTING)