Enable additional functionality in test program
Enable better printing output for logging.
Use argparse to accept arguments including -d --debug, -c --cleanup.
Functions are logically complete (mostly) and some functionality like
deleting networks and instances works.
Networks and instances should not be created if they are already in the
network_ or server_list.
:bug: When deleting subnets and their associated networks, if the subnet
is NOT already attached to the router, an exception will be thrown
because the subnet has no interface on the router.
Testing loop mostly works, but instance and network creation still need
to be implemented.